job/requirements.txt
google-cloud-storage
gff3
